Include if else block to avoid keep generating bcrypt

+++ ops/scripts/start_portainer.sh
@@ -15,10 +15,15 @@ else
 	DOCKER_COMPOSE="docker-compose"
 fi
 
-#Generate bcrypt from password
-P_BCRYPT=$(docker run --rm httpd:2.4-alpine htpasswd -nbB admin $PORTAINER_PASSWORD | cut -d ":" -f 2 | sed -e 's/\$/\\\$/g')
-
-echo "PORTAINER_BCRYPT=$P_BCRYPT" >> .env
-
-# start portainer in detached mode and make sure volume are recreated (rather than use potential previous state that my be erroneous)
-$DOCKER_COMPOSE up --detach --renew-anon-volumes portainer
+if ! [ -z "${PORTAINER_BCRYPT+x}" ];then
+  echo "PORTAINER_BCRYPT value has been set in .env already"
+  # start portainer in detached mode and make sure volume are recreated (rather than use potential previous state that my be erroneous)
+  $DOCKER_COMPOSE up --detach --renew-anon-volumes portainer
+else
+  echo "PORTAINER_BCRYPT value is empty"
+  echo "Generate bcrypt from password"
+  P_BCRYPT=$(docker run --rm httpd:2.4-alpine htpasswd -nbB admin $PORTAINER_PASSWORD | cut -d ":" -f 2 | sed -e 's/\$/\\\$/g')
+  echo "PORTAINER_BCRYPT=$P_BCRYPT" >> .env
+  # start portainer in detached mode and make sure volume are recreated (rather than use potential previous state that my be erroneous)
+  $DOCKER_COMPOSE up --detach --renew-anon-volumes portainer
+fi
